requirements_local.txt
wheel>=0.37.1
bcrypt>=4.0.0
chardet>=3.0.4
charset-normalizer>=2.0.10
click>=8.1.3
decorator>=4.4.1
defusedxml>=0.6.0
distlib>=0.3.5
dnspython>=2.2.1
email-validator>=1.3.0
entrypoints>=0.3
filelock>=3.8.0
Flask>=2.2.2
Flask-Bcrypt>=1.0.1
Flask-Login>=0.6.2
Flask-Mail>=0.9.1
Flask-SQLAlchemy>=2.5.1
Flask-WTF>=1.0.1
itsdangerous>=2.0.1
jsonschema>=3.2.0
multiaddr>=0.0.9
multidict>=6.0.2
notebook>=6.0.3
Pillow>=10.0.0
platformdirs>=2.5.2
PyMySQL>=1.0.2
requests>=2.22.0
rlp>=2.0.1
SQLAlchemy>=1.4.41
testpath>=0.4.4
typing-extensions>=3.10.0.2
urllib3>=1.25.8
virtualenv>=20.16.3
wcwidth>=0.1.8
webencodings>=0.5.1
websockets>=9.1
Werkzeug>=2.2.2
WTForms>=3.0.1
zipp>=2.1.0
gunicorn>=19.3.0
python-dotenv>=0.21.1
flask-babel>=3.1.0
# flask-babelex>=0.9.4  # Commenté car conflit avec flask-babel
flask-sslify>=0.1.5
flask_sitemap>=0.4.0
openai>=1.3.0
python-slugify>=8.0.1
beautifulsoup4>=4.12.2
lxml>=4.9.3 